Add superuser support for auth API and config API.
This uses the is_superuser() call to grant access to GCP project owners
for the Config Service Admin API (used to configure the git repository)
and the Auth API for configuring groups. This is in support of
automating the deployment of a LUCI stack.

R=maruel@chromium.org, nodir@chromium.org, vadimsh@chromium.org
BUG=

Tested:
Verified creation of a new stack did not require group configuration in
the auth service, but could instead use GCP project IAM settings.
Review-Url: https://codereview.chromium.org/2956533003