Update Swarming bot's auth server implementation to support multiple accounts.
Each task now runs in LUCI_CONTEXT with two (or less) accounts available:
  * 'system' is available if the bot has 'system_service_account' configured in
    bots.cfg.
  * 'task' is available if the task has a service account associated with it.

The default account (the one being picked by task processes) is always 'task'.
If it's not defined, there'll be no default account at all. Task processes,
if they are aware of LUCI_CONTEXT, may switch to 'system' account. This will be
used for Isolate and CIPD authentication in run_isolated.py.

'none' and 'bot' are still the only supported values for service account names
(i.e the bot is still unable to make an RPC to Swarming to generate tokens). It
will be changed in the next CL.

This change is safe to deploy as long as all LUCI processes invoked by tasks
with enabled authentication are built from revisions more recent than:
https://github.com/luci/luci-py/commit/c93e6294fd78c661f2a28909e45000ffe80e2a1a
https://github.com/luci/luci-go/commit/b782c74dd7d045d41ae85e6a90680efa61d24cfa

There's absolutely no changes to tasks that do not use service accounts (i.e
almost all of them currently).

R=smut@google.com, dnj@chromium.org, iannucci@chromium.org
BUG=730878

Review-Url: https://codereview.chromium.org/2959143002