config_service: handle disappeared configs
If conifgs were imported from a ref, but then ref was deleted, the current
behavior is to leave the ConfigSet entity intact and record an ImportAttempt entity
with the a message "Could not load commit log".

This behavior causes a stale cq.cfg in a ref config if the ref was deleted.
The stale cq.cfg is returned by get_ref_configs and it causes CQ to keep on
processing it, although users intended to disable this instance of CQ.

This CL:
- makes ConfigSet.latest_revision attribute not-required and updates code
  that assumes that latest_revision is always present
- clears the attribute if the ref does not exist anymore
- optimizes get_config_sets: loading list of files loaded the latest revision
  twice

R=cwpayton@google.com
BUG=736922

Review-Url: https://codereview.chromium.org/2979743003