Bump number of outputs limit to 4096.
This safe because even if Cloud DB's entity size limit is '1Mb-4',
TaskProperties is stored as an ndb.LocalStructuredProperty(compressed=True).
This means the embedded struct is compressed via python's zlib.compress() and
the data is quite compressible, so the expectation is that TaskProperties can
safely be significantly larger 1Mb.

TBR=aludiwn@google.com
BUG=

Review-Url: https://codereview.chromium.org/2951853003