Introduce 'service_account' field in NewTaskRequest RPC message.
It replaces 'service_account_token' field, which is still accepted but will be
removed once all clients are updated not to use it.

Update comments and TODOs to reflect the new plan for implementing service
accounts support. In particular TaskRequest.service_account_token datastore
property still stays: it will eventually be generated on the fly when the task
is posted.

R=maruel@chromium.org
BUG=731847

Change-Id: I457bae6a6103751e6e303e77353395f248f5410f
Reviewed-on: https://chromium-review.googlesource.com/611823
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>