swarming: Add support for native Win10 symlink
This is annoying because we need to detect if running on Win10, which is already
fun due to appcompat shim. With this out of the way, add the flag to make it
work. This is only helpful when developer mode is enabled, but this is probably
the right call to make.

This code was manually tested on Windows 10 Pro 15063.

R=vadimsh@chromium.org
Bug: 755746
Change-Id: I5750a2ebe2941b71eb8ff5f9d03cecbfa2317b49
Reviewed-on: https://chromium-review.googlesource.com/639754
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>