tools/run_on_bots: Enable running a task in batches.
A mixture of running in parallel vs serial, allow running a task in a number of
batches at a time.  This is useful, for example, for network heavy tasks that
need to run on all the bots, but not all the bots at once.

Bug:
Change-Id: Ia165ee9e7aca6dba6b8c6ece57dbd9fc85b5d0a9
Reviewed-on: https://chromium-review.googlesource.com/648588
Commit-Queue: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>