swarming.py: Add --task-output-stdout option
Enables the ability to control where the stdout of a task ends up going.
The options are;

 * Don't download / get at all.
 * Only output to the json summary file.
 * Only output to the console.
 * Continue to output in all of the above (default).

At the moment when using swarming on the build bots we end up with;
 - stdout in swarming
 - stdout in the collect_task output
 - stdout inside the summary.json file

For Layout Tests where the stdout is *long* this makes recipes take
significant time in just parsing the summary.json output.

BUG=731095

Change-Id: I0fa7228a5be21820b3089306f6e9c8646983159c
Reviewed-on: https://chromium-review.googlesource.com/654020
Reviewed-by: Tim 'mithro' Ansell <tansell@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Tim 'mithro' Ansell <tansell@chromium.org>