Let the bot know its identity, as seen by the server.
This will later be used to figure out what email corresponds to tokens that
are handed out to tasks that specify 'service_account: bot'.

An alternative would be to try to resolve the token into email locally on the
bot, but this turns out to be more complicated (in particular error handling),
and more wasteful (the server knows the email anyway, let's just send it back
to the bot).

R=maruel@chromium.org
BUG=767730

Change-Id: If1488ca8f1bd5e011f9df0f4bf45380d801424af
Reviewed-on: https://chromium-review.googlesource.com/683321
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>