[config_service] Fix location of config sets not updating
The .git endings still remained in the location field of the config sets returned by the api despite reimporting them. It just returned seeing that the same revision was already processed before, so I added a simple boolean to basically force a new set to be added to datastore.

Bug:509672
Change-Id: I3c4f6441544e53b9cf0b11e2f20705cb0d40edfc
Reviewed-on: https://chromium-review.googlesource.com/726384
Commit-Queue: Mun Yong Jang <myjang@google.com>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>