swarming_bot: Work around STATUS_DLL_INIT_FAILED on Windows
This is not really a fix but just a way to reduce the effect of broken bots, as
rebooting was observed to "fix" the problem.

A form of resource exhaustion cause Windows bots to die when initializing a new
process and it results in BOT_DIED tasks.

R=vadimsh@chromium.org
Bug:766877
Change-Id: I2586f56507d633e0e0bfabba54cadbd733ee8e8f
Reviewed-on: https://chromium-review.googlesource.com/730890
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>