[config-service] Add cron job for metadata for quicker validation
For context, the bulk of the latency when querying the validate-config endpoint came from
requesting metadata from services (according to the waterfall chart in the trace of the logs).

Also, increases the TTL on the memcache for the metadata to 10 minutes to match the cron job
frequency.

Bug: 509672
Change-Id: Ib2da124986c242aed258304b086fdabc28e0bad6
Reviewed-on: https://chromium-review.googlesource.com/738827
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: Mun Yong Jang <myjang@google.com>