swarming: Update columns to include duration, pending_time, pool; reorder
Change the order of the time related columns to make them more useful.

The task duration and pending times are useful indicators to know if anything
went wrong.

The pool permits to quickly determine what team the task are relative to, and
incites the user to filter accordingly.

Remove user; it is somewhat useful but not enough to be in the default list.
R=kjlubick@chromium.org
Bug:
Change-Id: I4068baa43a5c5a4613cdf6e69e5fd4dcfff7846d
Reviewed-on: https://chromium-review.googlesource.com/762160
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Kevin Lubick <kjlubick@chromium.org>