swarming_bot: reorganize bot_main.py code layout
No functional change, only move code around so that related code is grouped
together with one exception: s/monitor_call/_monitor_call/g since it is not a
'public symbol'.

This is a bit of an annoying change but I think this will improve the long term
readability of this file, which is already fairly long. Each section could be in
its own source file if desired.

R=vadimsh@chromium.org
Bug:
Change-Id: Ic6adbf8fcd2256531e3de55d0c71f20e89f3739c
Reviewed-on: https://chromium-review.googlesource.com/766508
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>