swarming bot: ensure 'logs' exists before creating a file.
'logs' is already created by __main__.py, so most places are already covered.
But before starting task_runner, task_runner_stdout.log is created before the
process is started. At this point if the directory was deleted, it's too late
and the open() call throws an exception.

This is very rare but it did happen.

R=vadimsh@chromium.org
Bug:
Change-Id: Ia28435af6f4ef6fa168fd95a71d4b1b90f7ce0f3
Reviewed-on: https://chromium-review.googlesource.com/769357
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>