[swarming] Add env_prefixes TaskProperty.
This will allow swarming task triggerers to populate e.g. PATH with
CIPD binaries, and/or PYTHONPATH with subdirectories of the isolate,
etc.

This would immediately be used to replace the similar functionality in
kitchen, and to allow the 'swarming' recipe module in build.git to
include standard infra binaries (e.g. logdog, vpython) in $PATH for
triggered tasks.

R=maruel@chromium.org, nodir@chromium.org, tandrii@chromium.org, vadimsh@chromium.org

Bug: 786503
Change-Id: If50875a9cd32afd96fb59a24a0f0eb254dd98ed0
Reviewed-on: https://chromium-review.googlesource.com/777484
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>