swarming: add 'maintenance' hack
In theory we want a new bot state, but that will come with the proto rewrite of
the bot API. In the meantime, this simple change will permit users, who want
this, to disallow handing tasks to a bot without marking it as dead or
quarantined. Since no DB schema is updated, there's no way to monitor these for
now.

It's an "hack" only in the sense that BotInfo.composite is not yet updated and
the state is not surfaced in the APIs yet. Other than that, it's not more
"hackish" that state['quarantined'].

R=vadimsh@chromium.org
Bug: 776832
Change-Id: I45305192ad9edd74c895e802415f26d37caec152
Reviewed-on: https://chromium-review.googlesource.com/802054
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>