Add dimension inside_docker when on linux
- inside_docker is set to 1 if '/.docker_env' exists, 0 otherwise. It can also
  be set to 'stock' or 'nvidia'. nvidia-docker detection will have to be added
  later by someone who has access to it.
- Clean up expectations in os_utilities_test.py a bit.

There's currently no way to detect from inside the container what is the runtime
version of docker, which would have been great to expose.

Explicitly do not use 'docker' as the key, since it could be incorrectly guessed
as to mean that docker is available. 'inside_docker' means that the worker
itself is inside docker.

R=vadimsh@chromium.org
Bug:
Change-Id: I194a0fb31d89561246aeac4860ce87c2d1ddad38
Reviewed-on: https://chromium-review.googlesource.com/806615
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>