Refactor run_isolated.py run_tha_test() to be more maintainable
18 arguments is too much.

There is a small functional change where if an isolated didn't have a command
specified, the arguments after -- would be automatically 'upgraded' from
extra_args to command. Disable this and require --raw-cmd to be used.

This reduces the risk of transparent failure.

This refactoring is done in preparation of issue 792491.

R=vadimsh@chromium.org
Bug:
Change-Id: I76adb3b9e2592b47b2e5b00053683d2d26903197
Reviewed-on: https://chromium-review.googlesource.com/814235
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>