Make 'maintenance' a monitored status in ts_mon_metrics
This maintenance field is based on the value of the 'maintenance' key in
the bot_info state dictionary, which will be set in the on_before_poll
method of bot_configs trying to prevent maintenance tasks from running
alongside swarming tasks.

maruel@ has already submitted a CL that makes it so that setting the
'maintenance' key to True treats the bot as if it were quarantined,
preventing real tasks from being scheduled on it.

Read more at http://bit.ly/2A5aAg9.

Bug: 776832, 416072
Change-Id: I24bbb9853f11a084829354db950ccef7a31f3954
Reviewed-on: https://chromium-review.googlesource.com/817986
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Charlie Andrews <charliea@chromium.org>