[auth] Make 'list_group' also return globs and nested groups.
Will be useful in UI.

Also make it return lists, not set, since the listing must not be used for
membership checks! Sets may give wrong impression that it is OK to recursive
fetch the entire group tree and then do checks against it. It is not.

R=tandrii@chromium.org
BUG=793516

Change-Id: I8d4ec6fe7283b3ba88a4a9c0a111d7c67254b619
Reviewed-on: https://chromium-review.googlesource.com/818571
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>