[auth] Add REST API for looking up nesting and owned groups.
Give a principal (an identity string, an identity glob or a group name), it
returns a group subgraph with all groups that include the given principal
(perhaps indirectly), and all groups that are owned by the given principal
(also perhaps indirectly, e.g if A owns B, and B owns C, then A indirectly
owns C).

The data is returned as a DAG where nodes are groups and edges are labeled with
'IN' or 'OWNS' relations. API users can traverse this graph to gather the
information they need. For example, to get all groups a principal belongs to,
one can traverse the graph along 'IN' edges from the root (the node that
indicates the principal) to all leafs.

R=tandrii@chromium.org

Change-Id: Ib8f3029ecfdc3cb9d6aa0cf4c3de89411e314563
Reviewed-on: https://chromium-review.googlesource.com/821418
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>