swarming: work around memcache woes
NDB has an extremely rare failure mode where a DB transaction succeeds but
the corresponding memcache cached entry is not updated correctly. In the
case where this happens at the very last write, where the task is marked as
completed but the memcache version stay stale, the task is returned as still
running by the APIs as the stale version of the entity is cached in memcache and
retrieved from there.

The work around is to not use memcache but only for the task result, and only
for an RPC fetching a single task result.

Change the behavior of the following task endpoints:
- cancel: doesn't fetch the task result anymore, it was unnecessary
- request: doesn't fetch the task result anymore, it was unnecessary
- result: only use memcache data if the task is completed, else fetch from DB

R=vadimsh@chromium.org
Bug:
Change-Id: Icd3b755954dde33328fdb1031084d9f3b1ab6b04
Reviewed-on: https://chromium-review.googlesource.com/836754
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>