Swarming bots.cfg: do not allow exactly equal bot_id and bot_id_prefix.
There is no reason to have the same bot_id and bot_id_prefix in the same
group:

    bot_group {
      bot_id: "one"
      bot_id_prefix: "one"  // not allowed and redundant.
      ...
    }

and similarly across groups it creates ambiguity:

    bot_group {
      bot_id_prefix: "some"
    }
    bot_group {
      bot_id: "some" // used to take precedence over prefix match above.
    }

This CL is backwards compatible on the server side in that old configs
will continue working as is, but new invalid configs will be rejected.

A follow up CL will change behavior internall as well.

R=vadimsh@chromium.org

Bug: 781087
Change-Id: Id1cf4cfc3208e347a7dcfd0a328cf9e7f1dd09e1
Reviewed-on: https://chromium-review.googlesource.com/849114
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Andrii Shyshkalov <tandrii@chromium.org>