[gae.py] Consider untracked files as tainting.
When uploading a new version files that are present but are not tracked
by git may still be bundled and uploaded, this makes the absence of the
"tainted" label in the version name incorrect.

With this change, any version that is uploaded when untracked files
exist in the checkout will be labeled as "tainted" the same way that
it currently works for uncommitted changes.

R=iannucci, maruel,stgao

Bug: 807384
Change-Id: I29a4381669545b3c864709feedadc3de1c0e05c1
Reviewed-on: https://chromium-review.googlesource.com/894185
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>