swarming_bot: add server_version as *forced dimension*.
It intentionally cannot be mutated by the bot hooks, which makes it a kind of
one off, but I think it's worth it.

This will be useful for live canarying, as someone could request bots running on
a specific version of the server.

It'll be cool in /botlist, as we'll be able to see which server version each bot
uses.

Update get_config() to handle BadZipfile transparently instead of crashing.

R=vadimsh@chromium.org
Bug: 812965
Change-Id: Id69bc58243990ac1003d19b44b0842ec6a582710
Reviewed-on: https://chromium-review.googlesource.com/923527
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>