[swarming] Correct fix_python_path to simply insert python.
Now that we frequently deploy python with the task, and no longer
have depot_tools in $PATH, it's better to actually invoke `python`
found in $PATH.

This implementation scans $PATH to find either `python` or
`python.exe`. If none is found, this falls back to sys.executable.

R=maruel@chromium.org, vadimsh@chromium.org

Bug: 804174,812693
Change-Id: If4827c5ba510358a8bfd6d8649cfb0117a13c59d
Reviewed-on: https://chromium-review.googlesource.com/933361
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>