[swarming] Relax duration and exit_code in task result
It used to require both to be set or not. Relax this rule, as this happens in
case of internal failure. Having this rule just causes even more trouble.

Condense the unit test a bit by created default expectations and then overriding
the expected difference. This highlights the difference more and shorten the
code.

This will help reduce the number of BOT_DIED caused by improper TIMED_OUT
handling.

Bug: 812711
Change-Id: I2b0d55f4b79732bb42db7bc0745759460755ffcb
Reviewed-on: https://chromium-review.googlesource.com/978381
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>