[swarming] Add TaskRequest.manual_tags and prepare for TaskRequest.task_slices
TaskRequest.manual_tags is to differentiate between tags specified by the user
and tags that are automatically added. Tags on a TaskSlice will have the manual
tags plus the tags relevant for this specific TaskSlice. E.g. if someone
triggers first on 'gpu:foo:bar' then fallbacks on 'gpu:foo:baz', the
TaskResultSummary shall have the automatic tag from the dimension of the
TaskSlice that actually ran. Said in a different way: this means that the tags
for a TaskResultSummary shall have only the relevant tags, hence the manual tags
must be kept in a separate property in TaskRequest.

Stop using TaskRequest.properties and use TaskRequest.task_slice(index) to
prepare for TaskRequest.task_slices. Right now index must be 0, as there's only
one TaskSlice (the one stored as TaskRequest.properties). This makes later CL
much simpler.

Redo task_request_test.py TaskRequest entity creation flow as this was based on
a flow that was used in 2015.

No behavior change from the user's PoV.

Bug: 781021
Change-Id: If4e49e577ef9086cb1560e0b19c54070174c738e
Reviewed-on: https://chromium-review.googlesource.com/995621
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>