swarming: fix stale index causing exception in tx
Between the time the BotInfo query yeild entities and the transaction runs, the
BotInfo entity could be deleted. This happens for Machine Provider managed
machines, as BotInfo is deleted once the lease is released, the bot is deleted.

This causes exceptions in cron_update_bot_info(), which is annoying but not a
big deal.

R=iannucci@chromium.org

Bug: 826421
Change-Id: I04791b34f9a8f4080cb91fa27927b19ae1d4d7b6
Reviewed-on: https://chromium-review.googlesource.com/1005281
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>