[swarming] reduce parallelism in cron_update_bot_info
A lot of CommitError are being raised due to contention on the production
server.

Lower from 25 parallel transaction down to 5 to try to help reduce the failure
rate of this cron job.

Put the code inside a try/finally so it can log how many items it processed even
when failing.

Bug: 826421
Change-Id: I647dbe3bcaffb9c704ab6a26f464315717b24329
Reviewed-on: https://chromium-review.googlesource.com/1006095
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>