[swarming] Retry Tx more in /internal/cron/update_bot_info
datastore_utils.transaction_async() sets the default number of retries to 1,
which means a lot of
  TransactionFailedError: too much contention on these datastore entities.
  please try again.
are thrown.

This is a problem to the extent that the logs are filed with exceptions due to
this cron job at a rate of ~50%, depending on the new number of dead bots.
'Fix' the problem by retrying more.

TBR=iannucci@chromium.org

Bug: 781021
Change-Id: I2a93b3380f17ca6c6e975df97a99b2832094c540
Reviewed-on: https://chromium-review.googlesource.com/1007334
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>