[swarming] variable aliasing strikes again
'lambda: foo(bar.biz)' will not work correctly when inside a loop. This resulted
in the same bot being run in the transaction. :( It took me an unreasonable
amount of time to figure this out.

Log the bots that are updated as dead.
Reduce log level otherwise to debug.

TBR=iannucci@chromium.org
Bug: 826421
Change-Id: Iada8e800e7006742bf074bc140a513202a12d831
Reviewed-on: https://chromium-review.googlesource.com/1012742
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>