[client] Add swarming.py post
This command permits arbitrary post RPC. Think of a generic RPC CLI tool, except
that it prepends the swarming API base resource.

While not generally meant be used, it'll be useful to test TaskSlice support
while iterating on the way this should be presented to the user in the CLI tool.

Bug: 781021
Change-Id: Ie405178e9b4849d69751d67e24eb17f486075edc
Reviewed-on: https://chromium-review.googlesource.com/1028461
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>