[swarming] Add queues/list RPC to list TaskQueues
This will be needed for monitoring. In the meantime, this can be queried with:
    ./swarming.py query -S chromium-swarm-dev.appspot.com --limit 0 queues/list

which is useful enough for my immediate needs. This will be a building block for
a new UI. I decided to create a new API namespace because I think I will want to
add more RPCs to it.

Bug: 803450
Change-Id: Ib28f0cc52b15308e7bfb337fb1160281f4f9f7a0
Reviewed-on: https://chromium-review.googlesource.com/1030790
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>