[swarming] dedupe non-zero TaskSlice
- When looking for task deduplication, also look for task slice fallback. This
  should be fairly fast as we don't expect task requests to have many idempotent
  task fallbacks.
- Improve task_scheduler_test to be more specific when looking at TaskToRun
  entities than "the one that happens to be fetched by the query".
- Remove spurious log statements.
- Change the cron job to reenqueue task fallback to return the task ids
  affected, this will be leveraged in the unit test that confirms this to work.

This change is currently a no-op as only a single TaskSlice is allowed. A unit
test is already written for this specific use case but I can't commit it right
away. :/

I still prefer to fix these bugs before, so once task fallback is enabled, the
code is actually known to be good.

Bug: 781021
Change-Id: Ie5943924a782994c717c4cb995ee2a9e07748b72
Reviewed-on: https://chromium-review.googlesource.com/1033443
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>