[swarming] retry Tx more in expired cron job
The expired tasks may throw a lot as there's concurrency while the cron job and
polling handlers are running. Retry the transaction more often in the cron job
to give it higher chance of succeeding.

This will hopefully reduce the errors a bit in the server log. They are not a
significant problem but just a tad noisy.

Change-Id: Icddcdc128bcc31ba68324b654a2974fb55b28133
Reviewed-on: https://chromium-review.googlesource.com/1045785
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>