[swarming] increase update_bot_info retries to 5 and hide transaction failures
The BotInfo entity is somehow very slow to update. This is weird because this is
happening for dead bots, which in practice means that the entity group has been
idle for 10 minutes. ¯\_(ツ)_/¯

Trap the CommitError exception, so that the cron job returns HTTP 200 even if
some entities were not updated. This is fine because it is run every minute, so
that it will be caught on the next run, 1 minute later. This needs to be
silenced because the failure rate becomes high enough it makes it fail the
canary, which blocks the deployment.

Change-Id: If4c9f7c80ff7fd58fcc6fa07ba887bf2964b7176
Reviewed-on: https://chromium-review.googlesource.com/1050105
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>