[client] better handle tarfile created on other OSes
When a tarfile is created, the paths are with native path separator. Sometimes
we may want to fetch the input on another OS (e.g. Windows files on a linux
host) for inspection, and this would create files with a '\' in their name
instead of a path separator.

Bug: 841878
Change-Id: I1352ae1fde59fcbbe40720375c9f340721239156
Reviewed-on: https://chromium-review.googlesource.com/1057424
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>