[swarming] Add capacity logic but still return True all the time
This extensively leverages memcache to determine if a task queue is valid or
not. My little pinky tells me there's an high chance this will blow up, so make
it always return True for now, and log if it thinks it would return False.

This will permit analysing the logs on prod, as staging will likely not be
useful, which will be useful to determine if the logic is working well under
load.

There are two different has_capacity functions, one in task_queues which checks
the task queue based cache, and the other in bot_management which does a BotInfo
DB query.

Bug: 839173
Change-Id: Iacaf13ac017fb433c27f477324b468018b7d6a38
Reviewed-on: https://chromium-review.googlesource.com/1050201
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>