[swarming] Update LKGBC in edge case of inactive bots.
I'm writing this CL because I'm tired of manually updating idle bots that
broken by the sheer fact of their LKGBC being over a year old.

Add a formal LKGBC section in Bot.md so it is easier to link to.

Includes unit tests for both the normal LKGBC update code path (which was
missing, oops) and the new one.

Bug: 843560
Change-Id: I235660a174adc19957b0423e5aadab40c05af455
Reviewed-on: https://chromium-review.googlesource.com/1061815
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>