[swarming] Change delete_old_bot_events to run every 5 min for 4.5 min
It was configured to run for 9.5 minutes every 10 minutes unsynchronized.
The 9.5 minutes run caused 'Exceeded soft private memory limit of 512 MB with
512 MB' by deleting ~30k entities per run, even if the handler is designed to
use as little memory as possible.

The cron job seems to be able to delete around 60 events/seconds, which is
likely not enough to keep up but that's still better than just accumulating.
Make it synchronized to it runs more for the allocated time. A follow up will be
to delete in pipeline more but the immediate need is to not polute the logs with
error messages.

Bug: 826343
Change-Id: I71cc2acf5c56e3d6b4861dc375240c4e1fc6f3d7
Reviewed-on: https://chromium-review.googlesource.com/1064723
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>