[swarming] capacity: move hardcoded True to _FAKE_CAPACITY global
This permits mocking the value at the test suite level, which enables testing
that is closer to the final expected result. It's now close enough O(days) to
being made live that testing the expectation instead of reality is now
preferable.

While updating test cases, I found out a bug in bot_reap_task() which I
commented. I decided to not fix the bug to keep this CL focused, and the bug
will be updated in a relatively simple follow up.

Bug: 839173
Change-Id: I8c05dc6dea578e26d9006725bc555378615424bc
Reviewed-on: https://chromium-review.googlesource.com/1070516
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>