[swarming] add wait_for_capacity to have task sit idle
There is a use case for having task wait even if there is no known capacity
after all. Make NO_RESOURCE the default, as we want people to use the flag if
they want their task to hang pending when there's no capacity but not encourage
people to do this.

The flag wait_for_capacity is member of TaskSlice, not TaskRequest, thus each
individual task slice can hang or not, based on the client preference.

Note that NO_RESOURCE is still in alert-only mode, so even if this code path
checks for capacity, in practice the task will *still* be PENDING instead of
aborted with NO_RESOURCE until the bit is flipped.

Bug: 839173
Change-Id: Ie7fc456c5b8945adc4f813537ccf1ba35fd306d7
Reviewed-on: https://chromium-review.googlesource.com/1070657
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>