[client] Update lru.py and test
Add iteritems() and transform().
Remove keys_set(), wasn't necessary. set(lru_dict) works as well.
Rewrite the unit test to be more extensive.

This is in preparation for named cache refactoring. transform() will be used to
upgrade the named cache format.

iteritems() is not strictly needed but is helpful for the tests.

get_timestamp() will be removed later on but keeping this CL simpler for now.

Bug: 844151
Change-Id: I3ff28d0c19551913fa5327cc7a8a26e806214142
Reviewed-on: https://chromium-review.googlesource.com/1100341
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>