[ui] Integrate JSDoc as a build rule.
Additionally, refactor swarming-app to use SwarmingAppBoilerplate, which
will be used by bot-list and every other top level app.

An update to elements-sk@2.1 lets us use the renamed and smaller
icon-sk elements.

This also re-works the imports to align with the best-practices,
that is, a module's index.js only be ./module.js and ./module.s?css
and all other imports to go in ./module.js

I found a few typos in the main ui/ page so fixes for those are
here too.


Bug: 842181
Change-Id: I59d0076e3ffc3b62717749d630f26f3459827259
Reviewed-on: https://chromium-review.googlesource.com/1108079
Reviewed-by: Joe Gregorio <jcgregorio@chromium.org>