[client] Implement the bot side of named cache size hint
The trimming code itself is still disabled, so there is no behavior change yet.

This is a breaking change for run_isolated.py: it's --named-cache argument now
accepts three parameters instead of 2, the third one is the named cache size
hint, or -1.

When run_isolated starts and it detects that some named caches are missing, it
makes sure that the free disk space is increased by the hint requested. This
code is not yet activated.

This means that a trim may occur before the task starts, if run_isolated.py
determines that more free disk space is needed.

Change the API to pass strings, so that JSON encoding doesn't get in the way.

Bug: 873736
Change-Id: I9b8a8837dba7ae4e12b3761e32c1cddcfb4a9ee8
Reviewed-on: https://chromium-review.googlesource.com/1177902
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>