Broadcast multiple device_types if necessary
This is to specifically fix the 2017 NVidia Shields (darcy)
that are currently broadcast as foster (2015 model).

A spot check of the devices on the fleet showed that Google-brand tends
to have all 3 values be the same and third party can have a wide variety.

Adds some tests that emulate real-world data.

Change-Id: Ic738fea9dc82f8d393fa54595b8e4287f3a7ba18
Reviewed-on: https://chromium-review.googlesource.com/1193762
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Kevin Lubick <kjlubick@chromium.org>