[swarming] Trim BotEvent older than 2 years
Used to be 3 years. Now that the backlog is done, it's fine to trim
more. We don't expect history older than 2 years to be of much value. This will
permit deleting hundreds of millions of entities.

Improve unit test to confirm that the right instance was deleted,
instead of just counting the number of entities.

Reorganize the cron code a little bit to keep less code inside the try/finally.

Stop using BotEvents as plural in the comments.

Bug: 884579
Change-Id: I9107696127e9623c222d0b3fe58b436e70ea9d3c
Reviewed-on: https://chromium-review.googlesource.com/1227034
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>