swarming: plugin api: add slice_number cancellation reasons
slice_number is necessary for scheduler to inform swarming of which
slice should be run.

cancellation reasons are necessary for escheduler to inform swarming of
the cancellation of tasks that were cancelled for reasons other than
being preempted (for instance, because they were malformed and caused a
scheduler error).

BUG=chromium:899306
TEST=None

Change-Id: I10b617d1062c96404acc1c0664abc62ca8015756
Reviewed-on: https://chromium-review.googlesource.com/c/1338603
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>