[swarming] Move isolate and cipd server settings to pools.cfg.
Currently this only allows a 'global' configuration across all pools,
but it would be easy to extend it to allow for per-pool configuration
later (e.g. for canary purposes).

The goal here is to be able to tag tasks with a single
`swarming.pools.version` tag which represents the config version of
everything that affects the creation of the task. settings.cfg isn't
fetched atomically with pools.cfg, and the isolate/cipd settings are the
only directly-affects-the-created-task stuff in settings.cfg anyway.

R=maruel@chromium.org, vadimsh@chromium.org

Change-Id: Ice23b8a672d04c2b739a01c40285e40c4a7b705e
Reviewed-on: https://chromium-review.googlesource.com/c/1340620
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>