[client] Convert the upload to be pipelined; various performance tuning
Originally the full 'items' list had to be materialized before
uploading. upload_items() now starts the lookup as items are processed,
which will reduce the overall latency.

Comment out logging and asserts in deep functions to help with performance.
Intentionally leave the logging lines in, as this could be useful for follow up
debugging. I do expect a significant performance boost by this CL, but not on
the order of magnitude that tarring will do. But both will combine, so this is
for the best.

Mark many functions as private.

After this the tarring will be much more efficient.

Bug: 825418
Change-Id: I7a470ae96b2b93c46459cf96e919b814ba264baf
Reviewed-on: https://chromium-review.googlesource.com/c/1348810
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>