[client] Make TaskChannel usable as a generator
For a class to be usable as a python generator, it must have:
- __iter__()
- next()

I added __iter__() and renamed pull() to next().

This will be useful as a multithreaded python generator.

Splitting as a separate CL to make it easier to review.

Includes relevant unit test.

Bug: 825418
Change-Id: Ic2e8a288a991dc214dea59e9533010679d0fde21
Reviewed-on: https://chromium-review.googlesource.com/c/1349377
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>