[client] Further use generator during enumeration
Leverage that TaskChannel can now be used as a multithreaded python generator.

This permits to send inputs to /contains RPC before the full isolated is
rasterized, which helps with reducing overall latency.

This change, coupled with the 3 previous ones results in 10 seconds saving on a
Windows 7 VM on a 2658 files/639MiB dataset.

The tarring should live between expand_directories_and_symlinks() and
_directory_to_metadata().

Bug: 825418
Change-Id: I9321250ade2a3838a7009351fc06295ad1e23399
Reviewed-on: https://chromium-review.googlesource.com/c/1349376
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>