[swarming] Remove TaskOutput removal cron job
It is not needed, as TaskRequest entity group will be kept for only 18 months.

It is safe to remove the handler in the same CL than cron.yaml, as
uploading a version with the trimmed cron.yaml takes effect immediately,
even if the default version is not bumped.

Bug: 884579
Change-Id: Iebb2094d299ccce202e32b3778de782486c639aa
Reviewed-on: https://chromium-review.googlesource.com/c/1354365
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>