[swarming] Delete old bot; Log timestamp for entities being deleted
The logging will help to know if the cron job are keeping up with the
creation rate so that all stale entities are deleted within the expected
time frame. This is done for both TaskRequest and for BotEvent.

Deleting old bots is important with Machine Provider created VMs
otherwise cruft just accumulates. We want to keep a clean and tidy DB.
:)

The cron job will be added once this code is deployed everywhere.

Bug: 884579
Change-Id: I6fd0d7526446e7439273c9f706d14ead86817d7a
Reviewed-on: https://chromium-review.googlesource.com/c/1363252
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>