[swarming] Start accepting X-Luci-Gce-Vm-Token headers.
They are optional for now in a sense that if such header is present, but it
cannot be parsed or verified, it is just skipped.

During the migration bots will keep sending both X-Luci-Gce-Vm-Token and
X-Luci-Machine-Token headers. If X-Luci-Gce-Vm-Token breaks for some reason,
authentication will fallback to X-Luci-Machine-Token.

BUG=901936
R=maruel@chromium.org

Change-Id: I818f138c7f34b3bf05bf57368edaeeb8a0d55b1d
Reviewed-on: https://chromium-review.googlesource.com/c/1375015
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>