[swarming] Add TaskResult message and all necessary submessages
They are not used yet. The proto change is done in a separate CL because this is
already large enough.

This will be needed when streaming the task results to BigQuery. We expect the
use of the same format than the API.

Include a small crash fix in TaskRequest.to_proto().

Bug: 917106
Bug: 870723
Change-Id: Ifa3a88dbf66bcbc5b862cdb14124bec55a1ba4b2
Reviewed-on: https://chromium-review.googlesource.com/c/1379351
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>