[swarming-ui] Refactors and clean-ups of bot-list
prepares to lay the foundation of task-list

Makes a button properly hide on new bot-list when not logged in.

Deletes bit-rotted tests involving query params. I think the
issue had been caused in Chrome by pushing the state every render
and so Chrome would stop updating it (and even log a message in
the console).  Maybe now that we don't update the query params in
render, we could one day test that?

Bug: 842181
Change-Id: Ifb420fed5a011c8d167d90dcf8bb73e4c4382114
Reviewed-on: https://chromium-review.googlesource.com/c/1392324
Reviewed-by: Joe Gregorio <jcgregorio@chromium.org>
Commit-Queue: Kevin Lubick <kjlubick@chromium.org>