[swarming-ui] Build task-list UI from bot-list
This creates a large chunk of the task-list UI by more or less
copying-pasting the bot-list parts and adapting them for the new
variable names and data structure.

I figured it would be best to do this all at once, since it's mostly
the same, and then work on the the task-list specific things like
datepicking, task counts, and the rest of the API calls in future
CLs.

Bug: 842181
Change-Id: Ib5ec6067a5905f4deb324c7f0eb0901c699ad106
Reviewed-on: https://chromium-review.googlesource.com/c/1404014
Reviewed-by: Joe Gregorio <jcgregorio@chromium.org>
Commit-Queue: Kevin Lubick <kjlubick@chromium.org>