swarming_bot: rename fake_swarming, migrate to httpserver
fake_swarming was written a long time ago. Update it to reuse
httpserver, which works well. This enables reducing the amount of
duplicate code. Rename it to use the same naming convention than in
client/tests/.

Reorder the imports in task_runner_test.py to make it clearer what comes from
where.

This work is split out as this separate change, as the next one is to
update task_runner_test.py to use this, instead of mocked HTTP request
expectations, and this one will be fairly involved.

Bug: 921842
Change-Id: I85086428627287d2ddeb8587f59a541009d10d73
Reviewed-on: https://chromium-review.googlesource.com/c/1412258
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>