[swarming] use task queue to delete old tasks
There's just too many tasks created per day, so using a single thread is
not fast enough.

This change, once enabled, will significantly increase the cost of the
largest instance until the backlog is processed. If this is a problem,
the task queue can be paused.

Bug: 884579
Change-Id: I6eeb34707f30b0af1a53807c200698e7c136f4a7
Reviewed-on: https://chromium-review.googlesource.com/c/1434454
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>