[isolate] Remove invalid namespace historical values
These where never used in practice on the server.

There is a risk that clients interpreted these values. If this happens,
we'll have to update the clients, and possibly revert this change in the
meantime to not cause disruption.

Bug: 787113
Change-Id: I1bb979108c54fa88bb3750642aac1cdcccbb35b2
Reviewed-on: https://chromium-review.googlesource.com/c/1452348
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Auto-Submit: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>