[client] run_isolated closes file handles before starting task process
This changes the behavior of the Swarming bot to stop leaking processes
into the task.

This works *because* run_isolated.py doesn't redirect any of stdin, stdout nor
stderr; it's task_runner.py that has this responsibility. close_fds=True is not
supported on Windows when standard input or outputs are redirected.

Bug: 929965
Change-Id: Ibf338852aa838ff66964d035af1c449f8acf236f
Reviewed-on: https://chromium-review.googlesource.com/c/1459561
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>