[swarming] Change cron jobs URLs into manageable categories
All cron jobs have an URL starting with /internal/cron/ but after that,
it's a bit free for all. This CL aims at fixing that.

Put the cron jobs into subcategories to help readability and monitoring:
- cleanup/: low priority work to trim the DB; the service can work just
  fine without these but cruft will accumulate.
- monitoring/: important work to enable service monitoring; the service
  can work without these but we'll fly blind.
- important/: critical work to enable a functioning service, including
  scheduling

Then each category can have a subsection 'bots', 'tasks'.

This way we can alert on important cron job and ignore the noise on
lesser important ones.

This change is designed to be not disruptive, the new handlers are added
along the old ones, then a second CL will be created once prod is
deployed to switch over to the new handlers and remove the old ones.

Bug: 912154
Change-Id: I4e29df99aded8a2451fe816c8774f58c6ba92430
Reviewed-on: https://chromium-review.googlesource.com/c/1465957
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>