[swarming] Finally (!) fix a long standing bug in timeout handling
When a bot has trouble terminating a task due to a timeout, it may send
a task update with no duration. This would cause the server to kill the
task with an internal failure, leading to a BOT_DIED.

Fix this server-side, and add a unit test to confirm this works.

Add a note that this state (terminating, but not yet killed) will later be
properly stated as TERMINATING.

Change-Id: I82eb9a2dcc98cfd7c3664294663547fd6223d207
Bug: 916560
Reviewed-on: https://chromium-review.googlesource.com/c/1471319
Auto-Submit: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>