[swarming] Redo the backend Swarming URLs for tasks
Similar to the cron jobs, organize the task queues per importance.

Remove task queue 'cleanup', it isn't used anymore.

Do not rename the task queues for now, as this is more work, and '/' is not a
valid character in the task queue names, so '-' would have to be used, and it's
less nice.

Bug: 912154
Change-Id: Ibe783a124f1c4c689511f099668bf36499fbe193
Reviewed-on: https://chromium-review.googlesource.com/c/1470599
Auto-Submit: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>