[swarming_bot] Fix incorrect exception handling with win32com
com_error belongs to pythoncom, not win32com.client.

I don't have a nice way to test this, the code path only happens if COM triggers
an exception. so I intend to commit this while a bot is broken and see if it
helps handling the exception correctly. The bot is broken in the meantime
anyway.

Bug: 935568
Change-Id: If94aed00fe773d6a29c0dfbda940245b16e7065a
Reviewed-on: https://chromium-review.googlesource.com/c/1486752
Auto-Submit: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>