Adding fix_protobuf_package() to appengine_config.py of config_service
This means fix_protobuf_package() will always be run no matter which
handler is the first to get called.

Note that fix_protobuf_package() is idempotent, so it doesn't matter
if it gets called a second or third time.

Having said that, we should replicate this change to other apps
that call fix_protobuf_package(), and then remove the call from
any/all libs that call it, and update the lib's documentation to indicate
that fix_protobuf_package() should be added to appengine_config.py
as a prerequisite for using the lib


Bug: 928594
Change-Id: I94818802e0ff547451e98d3d964c7b6c88484cd5
Reviewed-on: https://chromium-review.googlesource.com/c/1496618
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>