[swarming] Backfill BigQuery for the past year (364 days)
Add link to the document as to why it's one year and not 1.5 year. Use
one day less (364) to reduce the risk of the last day being rejected
(with latencies in processing).

This will cause elevated App Engine frontend workload for a while (cost
on BigQuery side is minimal). There are two options that can be
simultaneously used if this is problematic budget wise:
- Reduce the backfill for each table to 6 months (182 days) and give up.
- Reduce the speed at which backfilling of each table occurs (30x real
  time at the moment); could be reduced to 15x or 10x.

Bug: 765435
Change-Id: Ia92207eec3b252016f35a9a5c22871f256b1fe06
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/1524001
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>