[swarming] Use a slightly les hackish sleep
Use standard ndb.Future.wait_any() and a list instead of a set of
ndb.Future. This should help with the deadlock observed on prod. Add a
log entry to confirm that the log is happening before or after, to see
if it helped or not.

Bug: 900042
Change-Id: I264e6002bfdfcf28e23611cfe5902fb839e5979d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/1534883
Commit-Queue: Jao-ke Chin-Lee <jchinlee@chromium.org>
Auto-Submit: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>