[swarming] Add support for offset in stdout
This removes the limit of 16MiB when retrieving data. There is still a
limit per RPC, but this can now be paginated.

Include the task result, so that the client can keep on polling stdout,
and when the state changes from RUNNING, the client can fetch the final
task results; or in the case of switching back to PENDING, can reset
since it's a retry.

Bug: 867621, 865069
Change-Id: Ie6c95ebf0653e9aaad22050274fe3aa887cc2d4b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/1542132
Commit-Queue: Jao-ke Chin-Lee <jchinlee@chromium.org>
Auto-Submit: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>