[swarming-ui] Break up large logs for performance reasons
This removes the quadratic reflow time.

I did not try breaking the logs up on every newline,
but that could be something to try in the future.
(I'm not optimistic about that approach, since then it
seems that we are trying to reflow the logs ourselves in
JS instead of letting smart browser developers do it in C++)

Bug: 867621
Change-Id: If5e3a5e42c493352653733bacceac178c57feae8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/1548742
Commit-Queue: Kevin Lubick <kjlubick@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>