[isolate] Throttle deletion more; 25 instead of 50
This happens on prod in /internal/taskqueue/cleanup/expired. We don't
want it to fail, so slower is better.

The failure mode is really stupid, we get either:
  raised RuntimeError(maximum recursion depth exceeded while getting
  the str of an object
or
  Deadlock in <Future ...>

which neither really say anything useful. Thanks ndb...

It is known that throttling the number of concurrent ndb.Future for RPCs
helps increasing the success rate.

To be clear, this is completely trial-and-error until the failure rate
becomes low enough to be acceptable.

R=jchinlee@chromium.org

Bug: 942065
Change-Id: I4bfccc10ed5e56dc5095d98862ad0830820de4c4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/1535672
Commit-Queue: Jao-ke Chin-Lee <jchinlee@chromium.org>
Auto-Submit: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>