[swarming] Silence Timeout in rebuild-task-cache
This can happen because the exception can be thrown when ndb.Future's
get_result() method is called, and it'd be too much trouble to handle
all the call sites. So catch it at the handler level to return a 429,
thus the handler will be safely retried without an alert (due to a 500).

R=jchinlee@chromium.org

Change-Id: I36e867b679dcb2f0253ab7137a42e410ee89fa4b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/1548741
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>