[isolate] Try harder to make the expired cron job succeed
This change is reactive of the behavior of the code on production.

The new code comes from the observation than before this change:
- /internal/cron/cleanup/trigger/expired failed most of the time, but not
  all the time.
- /internal/taskqueue/cleanup/query_expired fails occasionally, but not
  often.

And the only difference in the ContentEntry query between both is that
the second uses a two side boundary for the query while the first only
limited on one side.

So change the first to use two boundaries, we expect this to largely
increase the success rate.

Search 300 days ago instead of 360.

Bug: 942065
Change-Id: Ic318dbd1d1d90b397a975b7545b149a6511fbe96
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/1574259
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>