[swarming] Add Makefile in proto/; Rename proto package in config/
Add proto/Makefile to update all *_pb2.py files in all subdirectories
in one swoop.

Update package name:
- api stays as 'swarming.v1'.
- config used to be 'swarming', rename it to 'swarming.config'.

This makes imports saner. Config is not versioned since we can't
effectively version it. API is another story, and can be versioned. But
'swarming.api.v1' would be clunky and kinda weird for users.

Do not move the files yet into a v1 subdirectory, this will be done
later when we need to add a v2, in fact we could rename 'api' to 'v1'
and add a 'v2' aside.

Bug: 808836
Change-Id: I4bbd4bb32630cc67f8406f349b67b43d4f83baff
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/1622653
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>