[auth] Adjust SignedAuthDB to reuse existing serialized and signed data.
Also fix misleading comments in replication.py. "AuthDB blob" is actually
serialized ReplicationPushRequest message (which contains AuthDB along with
revision number and timestamp). This is exactly what we need to be signed,
and it is already being signed when it is pushed to replicas. We'll just reuse
that.

R=fmatenaar@chromium.org, maruel@chromium.org, nodir@chromium.org
BUG=959427

Change-Id: I6b53d1879f683398378110173681049bedbe20ea
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/1628178
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>