[swarming]: Note deprecated fields, log for clients using these
The goal is to remove these, but it cannot be done until clients are
updated. Log an error when this happens, so we can ensure no clients are
using the deprecated fields. Once we know for sure, we can delete the
deprecated fields.

No functional change, beside one additional log.

R=jchinlee@chromium.org

Change-Id: Id4cf2e780cf72c11b1ebcb68d09fa54f4443e022
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/1641310
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>