swarming: pull legacy TaskState enum into a proto
Rationale: the legacy json rpc is still the main supported interface for
making task state queries into swarming. It's definition of the
TaskState enum is unique to that API (not shared with the forthcoming
proto rpc definition), so re-using the enum definition from clients
requires duplicating the enum definition in those clients.

Instead, provide a proto representation of this enum that clients can
use.

BUG=971376
TEST=None

Change-Id: I3037fc5a167ae09b93683dafad6d93f2a08ab39b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/1653718
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>