[auth] Put a limit on number of GCS ACL entries.
The hard limit imposed by GCS is 100 ACL entries. When it is reached, AuthDB
updates would stop (because GCS writes would fail). This is bad.

We prevent this by checking the number of existing authorized readers when
registering a new reader. Better to reject a single new reader than break
everything by going over the limit.

What to do when the limit is actually reached is a problem for another day.

R=maruel@chromium.org
BUG=959427

Change-Id: I9d607b71e86e8eefd6be059d10a0949b4e8880c9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/1763053
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>