Find arg0 via $PATH for non-Python cases
Splits fix_python_cmd into two methods, one for appending "python" to
the front of a command if running a .py file, and one for finding arg0
via $PATH.

This is so that executables that may be passed for arg0 can still be
found via $PATH, e.g. "python3" or "vpython".

Bug: 898348
Change-Id: Ibab11929b446bd96f19f86285278ef33c21dbb98
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/1779153
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>