[swarming] Disable logging for /bot_code
That's the only case where it is triggering on the -dev instance. I
suspect it is going to be overwhelming on prod, which will hide the data
I actually care about, so hide this log for now, as it can be handled
later.

No significant change beside less logging.

R=tikuta@chromium.org

Bug: 1010555
Change-Id: I9d6e937656112c68b86b2c2d50ae36de93c86117
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/1847908
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>