[swarming] remove local function
This is a follow up for
https://crrev.com/c/1916984/15/appengine/swarming/server/bot_management.py#727

I thought separating the sub task is readable.
But it creates a local function unnecessarily, and isn't not readable for some people.
Keeping them at the same place would be easier to see in this case.

Bug: 916578
Change-Id: I23ca87bc78635008c4325f0da278b1f5505fc6d5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/1935528
Auto-Submit: Junji Watanabe <jwata@google.com>
Commit-Queue: Takuto Ikuta <tikuta@chromium.org>
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>