[swarming] external scheduler: do not raise exception when reaping nonpending task
When external schedler is assigning a task, the task is sometimes not pending.
The task is either running or finished, including failures.

In that situation, there is nothing swarming can do other than notifying to external scheduler
to fix the inconsistent state, and the exception can be ignore.

This is a dominant error in chromeos-swarming console.
http://screen/AJ1tsMNA7DZ

Bug: 1032894, 1027448
Change-Id: I434a32ad7e775d662a461f35892802798310058c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/2006998
Commit-Queue: Junji Watanabe <jwata@google.com>
Auto-Submit: Junji Watanabe <jwata@google.com>
Reviewed-by: Xinan Lin <linxinan@chromium.org>
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>