[Config] Upgrade Config UI to Polymer 3.0 for HTML Import Deprecation
Background: Chrome will stop supporting HTML import since M80[1]. The
current config UI use this feature heavily because it is the recommended
way for component packaging when it was built. As of today, ES Modules
is the recommended way moving forward and adopted in Polymer 3.0
(currently, the UI is on 2.0). This change is to move the ui to polymer
3.0 and replace the usage of HTML import with ES Modules.

Most of the changes are following this instruction[2] to turn HTML
import to module import and switch from bower to npm. Files are moved to
the place where makes more sense to make the file structure clear.
There're also minor changes listed below needed to adopt the breaking
change in Polymer 3.0

1. call `super.ready()`[3] in `ready` callback before accessing DOM
2. use mixin for legacy beahviors[4]
3. fire api is gone. Replace with dispatchEvent API[5]
4. Remove the copy of google-sign-in element from third party library
  and consume it from npm[6] as it is now compatible with Polymer 3.0
5. `async` api is gone. Use native setTimeout instead[7]

[1]: https://www.chromestatus.com/feature/5144752345317376
[2]: https://polymer-library.polymer-project.org/3.0/docs/upgrade
[3]: https://polymer-library.polymer-project.org/3.0/docs/devguide/custom-elements#ready-callback
[4]: https://polymer-library.polymer-project.org/3.0/docs/devguide/registering-elements#using-legacy-behaviors-with-class-style-elements
[5]: https://polymer-library.polymer-project.org/3.0/docs/devguide/events#custom-events
[6]: https://www.npmjs.com/package/@google-web-components/google-signin
[7]: https://polymer-library.polymer-project.org/2.0/docs/upgrade#async

R=tandrii, vadimsh

Bug: 1038927
Change-Id: I3f53db13553ab75deb7735a4eca6bc7c13b4b2f5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/2021339
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>