[realms] Use uint32s to reference permissions in Binding message.
This is for the internal representation only.

I realized that due to flattening of all realm configs, permission
strings in Binding.permissions will be repeated A LOT. This will
significantly balloon the size of Realms message. It will likely
compress well, but we sill will have to keep it in RAM in an
uncompressed form to work with it, so it still makes sense to do
some easy memory optimizations.

R=iannucci@chromium.org, nodir@chromium.org
BUG=1051724

Change-Id: Ifcb8dc954700f7e90c3e741b38e62a574c52baf5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/2062908
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>