[realms] Implement ingestion of the permissions list.
The list of known permissions now lives in the AuthDB entity group in
AuthRealmsGlobals entity (named that way in case we need to put more
stuff into it).

It is a true part of the AuthDB, hooked up to all regular places such
as proto<->entity conversions and the change log.

R=iannucci@chromium.org, tandrii@chromium.org
BUG=1051724

Change-Id: I8d81f12ad9d61c0b49e5b77ed22be985e0646478
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/2092844
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>