[realms] Implement expansion of realms.cfg into a flat form.
There are many ways to combine what is conceptually a list of
(principal, permission) pairs into a list of Bindings (which are
([principal], [permission]) pairs). It's not clear what is the best
one among them.

This implementation tries to "merge" role assignments when possible,
e.g. if a principal1 is OWNER and principal2 is READER+WRITER and
OWNER happens to be == READER+WRITER, then principal1 and principal2
will end up in the same Binding (since their factual permissions are
identical).

Also this implementation tries to preserve memory by doing more
computations instead. We assume the realm inheritance graph is
relatively shallow.

R=tandrii@chromium.org, iannucci@chromium.org
BUG=1051724

Change-Id: If442566f47fb14a3817e2e9ede2745d13dbc4127
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/2096142
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>