swarming: Don't apply task templates to retried tasks.
Currently, if you retry a task that used chromium's task templates,
the server will deny the request with an error like:
"request.env_prefixes[u'PATH'] conflicts with pool's template"

This is because we try applying the template to a task that already
had the template applied. And since a template conflicts with itself,
the retry request fails.

This fixes that by adding the special 'task_request.TEMPLATE_SKIP' flag
to tasks launched via the retry & debug buttons.

Bug: 812428, 1017740, 960835
Change-Id: Ief398332f27ed30aed9c91ac132cd62015ef25a7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/2097048
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Reviewed-by: Junji Watanabe <jwata@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>