[auth] Make replicas always use load_sharded_auth_db.
load_sharded_auth_db is a code path that loads all Auth state from
a single proto message stored as a compressed blob in the datastore.
It has been the default code path for ~5 months, but there's still
a fallback to loading AuthDB from individual entities. This CL
removes the fallback.

It would allow to remove the code path that "explodes" an AuthDB
proto into a bunch of entities when replicas receive a push. This
would somewhat simplify Realms implementation.

Note that we still need code to construct AuthDB from entities, since
this is what's being used on auth_service or when services are running
in standalone mode.

This CL also includes refactoring of AuthDB class to not store Auth
entities internally. Constructing them is just a waste of time and
memory when AuthDB instance is built from a proto message.

R=iannucci@chromium.org, tandrii@chromium.org
BUG=1051724

Change-Id: Ib755bd7036a21b7a61fd1a87501ae7eaeb5dfc86
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/2099452
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>