[auth] Fix botched refactoring.
model.root_key is a function, we need whatever it returns, which
happens to be available already as 'root_key' local war.

Caught by Buildbucket tests that wipe mocked datastore. This
condition should be impossible in real applications (and there was
no test for it).

TBR=tandrii@chromium.org, iannucci@chromium.org
BUG=1051724

Change-Id: I3621c26f0c29b9c4df81d59db27ddbfcf050ae01
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/2106900
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>