Re-cache free disk space when _load fails
The `DiskContentAddressedCache` calculates free disk space from the size
of the `cache_dir` on init, and then possibly again on `_trim()`.
However, when both of these cases are true:

1. The LRUDict fails to load (for instance in the event of a version
change)
2. Trim on init is false (which is the case in `run_isolated.py`)

then the cache is deleted, but free disk space is not recalculated,
resulting in the cache thinking it's using more disk than it actually
is.

To address this issue, recalculate free disk space when the cache load
fails.

Bug: 1062859
Change-Id: I6b4570af9512688290540bf3b0f5aaeca4fda90d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/2109276
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Commit-Queue: Takuto Ikuta <tikuta@chromium.org>