[realms] Register Buildbucket permissions.
They are not really used yet, but already instantiated in
Go and Python code bases, causing the auth library to warn
on startup that they are not defined in the realms DB. This CL
should remove the warning.

Also register `buildbucket.builds.update` while at it, it will
be used as a replacement for `buildbucket-update-build-users` group.

R=smut@google.com, nodir@chromium.org
BUG=1091604

Change-Id: I2f702fc5483b7cf533cd47626a4c115d4fd6bba9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/2300538
Reviewed-by: smut <smut@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>