[auth] remove role/swarming.pool.Viewer from role/swarming.pool.User
The main use cases of poolUser is to trigger tasks via led or swarming client. And go to task pages to see the result.
So no need to see task list or bot list.
It's easier to add it later if necessary than remove it.

Discussed in
https://crrev.com/c/2321371/4/appengine/auth_service/realms/permissions.py#114

Bug: 1066839
Change-Id: Idfdf557c586439653d941245edee2a3a2236c543
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/2328253
Auto-Submit: Junji Watanabe <jwata@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>