[swarming][realms] store 'realms_enabled' flag on TaskRequest.
When a task gets created in Realm mode, we want to always use Realm ACLs for the task.
But we don't know if the task was used in Realm mode or not after task creation.
This is problematic when checking user's authorization after task creation.
e.g. GET /request, POST /cancel etc

This CL stores 'realms_enabled' flag to TaskRequest entity to be able to retrieve the auth mode later.

This issue was originally discussed in https://crrev.com/c/2321733/2//COMMIT_MSG#11.

Bug: 1066839
Change-Id: I8160eeda91e5c9ccdf0ff9f769ae89adadd3e9e3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/2334827
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Commit-Queue: Junji Watanabe <jwata@google.com>