[swarming][qscheduler] Add tasks with parent, grand parent task.
This CL adds tests to reproduce ExternalSchedulerException we experienced in crbug.com/1127726.

The downstream CL, which caused the error, failed as expected.
https://logs.chromium.org/logs/infra/buildbucket/cr-buildbucket.appspot.com/8867836951502870992/+/u/swarming/run_tests_python2/stdout#L4251_1

The verbose logs of the downstream CL http://go/paste/5469601526710272 show that it failed with
 "BadRequestError(cross-groups transaction need to be explicitly specified (xg=True))".

This is same with the errors in the prod logs.
http://screen/ASVSuTBkbTXQAuW

Bug: 1126488, 1127726, 1131822
Change-Id: Ib16f1c1f20ed10bea5ba86d48ce1aba774e90982
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/2437658
Commit-Queue: Takuto Ikuta <tikuta@chromium.org>
Auto-Submit: Junji Watanabe <jwata@google.com>
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>