[subprocess42] Poll process in short intervals when waiting on windows
The current behavior on windows is waiting for the entire timeout.

Unfortunately, it looks like the process is not interruptible when
waiting on the subprocess it has launched (i.e. during proc.wait()).
Therefore, this cause the issue that the process is not able to act upon
a signal received while waiting until the full timeout duration is
exhausted or the launched process has returned. See experiment
in https://crrev.com/c/2530416

This change also makes the behavior on windows matches that on other
platforms.

R=jwata, maruel, tikuta

Bug: 1147338
Change-Id: I43296840f32b0c462f59f82b9adff80481aa1fea
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/2530538
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Reviewed-by: Junji Watanabe <jwata@google.com>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>