[luci-config] Add Team definitions to projects.cfg
The motivation here is to keep an up-to-date mapping of team
grouping/ownership of projects, as well as contact emails to
reach the humans of that team.

We often have need of this information when sending LUCI-related
emails, especially if there are action items associated with
that communication.

Currently we have this information (roughly) in a separate
spreadsheet which is sort-of-but-not-really up to date. The hope
is that by putting the information directly in projects.cfg we'll
be able to keep the information fresher, and project owners will
also be able to update the information as needed.

R=efoo, tandrii

Change-Id: Ib851e81c790144d6237dda8405da015481ea7e82
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/2893489
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Reviewed-by: Eric Foo <efoo@google.com>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>