[gae.py] Make sure app.yaml is in a subdirectory when staging.
When staging Go files for deployment cloudbuildhelper builds
a GOPATH tree, and then creates a symlink named as `dest` (from
the manifest YAML) that points to the GAE module directory in
the staged GOPATH.

This is problematic when `dest` is ${contextdir} itself. It already
exists, and it is a directory (the root of the staging area).

Make sure the manifest never contains '${contextdir}' as `dest` by
moving the ${inputsdir} root one level up (adjusting all manifest
entries accordingly).

R=iannucci@chromium.org, tandrii@chromium.org
BUG=1234651

Change-Id: Ieb3f2097fadfeac2121c13a01eac82282a54628a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/3064778
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>