client: Add nsjail containment flag to client
Adds the nsjail containment flag to run_isolated.py. This involved:
1. Modifying run_isolated command line args to include NSJAIL.
2. Adding an NSJAIL containment_type to subprocess42.py
3. Provisioning the nsjail package where necessary (and only on
   linux-amd64 bots)

Note that this CL does not fully enable using NSJAIL as a containment
type, as it does not properly set up the nsjail sandbox. This would
require modifying the task request, so we leave that to future CLs.

Bug: 1227833
Change-Id: I27e160ef3f5d11e970600699c7ff0c08d64056eb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/3064133
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>