swarming: List a GCE bot's networks in its state
Chrome's GCE project puts all its bots in a single network, which is
causing that network's quota to approach the limit. To reduce the
quota usage, we need to shard some bots across different networks.

Before doing so, it'd be helpful if a swarming bot can expose the
network(s) it's attached to. That way, spot-checking a bot's network is
painless.

So this adds a bot's networks (discoverable in GCE metadata) to a GCE
bot's state dict.

Bug: 1296250
Change-Id: I7f81f5059aa3c53adf0f10a46a6d8ff3688dbff1
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/3508313
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>