[auth] Setup a group for listing builders across *all* projects.
Will be used by Milo to cache builders list locally in a cron
job (I presume). Milo already has this access, but it needs to
make a per-project RPC (each one authenticated as "project:<name>").

Note that once permission.py is converted to a real configure file,
we should probably replace `buildbucket-builders-readers` with an
explicit list of accounts. But for now it is simpler to use a group,
since group name is the same in prod and dev universes (unlike Milo's
service account name).

`role/luci.internal.buildbucket.builders.reader` is marked internal
to avoid unnecessarily exposing very fine grain roles publicly. It
is much harder to delete a public role than to add it.

Also improve the test a bit to make it catch malformed bindings.

R=iannucci@chromium.org, weiweilin@google.com

Change-Id: I48fd159415a86980b3932d3b340f066caa9880fa
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/3502968
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Weiwei Lin <weiweilin@google.com>
Reviewed-by: Weiwei Lin <weiweilin@google.com>