[auth_service] Fix pylint errors.
Pylint was updated to 2.7 in depot_tools which caused previously
passing code to emit errors.

Note that 2.7 is leaning towards python3, but 90% of code in
luci-py is python2-compatible and will likely stay that way
until it is all decommissioned. For that reason disable
useless-object-inheritance lint globally in the repo (it is
about `class T(object)` which is meaningful statement in python2).

R=iannucci@chromium.org,
CC=justinluong@google.com, meiring@google.com

Change-Id: I110c0557f3e0ae9eb4fab79c23c7eb244dbcb4d0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/3783075
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>