[cert-factory] Updated luci-py deps according to tokenserver change.
We recently changed one field in the machine token and luci-py has a
dependency. Updating it here with the source of truth. Also realized
that the proto generator is very annoying! A lot can be done to it but
did bare minimum to make it work.

R=vadimsh@chromium.org

Bug:1197101
Change-Id: I4067b228d9748c1d0a224ac63b9228e1c81fd703
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/3774978
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Akash Mukherjee <akashmukherjee@google.com>