[swarming] Remove some dead code from bot events API.
It appears there was some unfinished migration or something. Remove
its remnants, we aren't planning to finish it now and bogus bot event
types just make code harder to understand.

An event type is considered unused if searching for it in the codebase
returns only hits in protos and unit tests (but not in the rest of
Swarming server code, nor in the bot code).

R=iannucci@chromium.org, jonahhooper@google.com

Change-Id: I18a65e52f7e928b9b632872b256ec3d77300d408
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/3821423
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>