[swarming] Implement the new way of getting active queue numbers.
Queues matched to a bot are stored in BotDimensionsMatches entity.
Every time the bot polls, we verify the corresponding queues are
still alive and still match the bot. This heavily relies on
ndb cache to be performant. But doing it this way removes the
complexity associated with "proactive" unmatching of bots from
queues in clean crons. If the bot doesn't poll, it doesn't really
matter that the list of queues assigned to it is stale.

If we detect bot changes dimensions we launch **an asynchronous**
task to find new matching queues (the task is not implemented yet).
In the previous implementation this was done synchronously, resulting
in frequent timeouts when bot had to rescan lots of queues in
/bot/poll.

In the new implementation, when bot changes dimensions, it will
immediately stop polling queues it doesn't match anymore (this
is required for correctness), but will start polling new queues
only some time later, when the TQ task completes. This is fine
though. In most cases bot dimensions changes are not resulting
in any **new** queues being polled (most notably, changes to
"server_version" dimensions do not really affect any queues).

R=jonahhooper@google.com, iannucci@chromium.org

Change-Id: If4a05c6340211f1b62eb5fbecf648e5d2ddf8927
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/3900212
Reviewed-by: Jonah Hooper <jonahhooper@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>