[swarming] Implement TQ task that finds bots matching tasks.
It is called whenever a never-seen-before kind of a Swarming
task appears.

Its purpose is to quickly find bots that can run this task and
"register" them to poll the queue with this new Swarming task,
so it is picked up ASAP.

If for whatever reason this mechanism is offline, bots will still
discover tasks, but only when they do full rescan of matching
tasks, which happens relatively infrequently.

R=jonahhooper@google.com, iannucci@chromium.org

Change-Id: I069588490f7b88d26882614ede8e4fcc4178bfe4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/3900837
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>