[swarming] Switch /queues API to use new entities.
As far as I know it is used only in UI to "auto-complete"
dimensions. As such, it is not very critical and we can switch
to new entities in advance of switching everything else to use
them (new entities are already kept up-to-date in datastore, even
though nothing is using them yet).

R=jonahhooper@google.com, iannucci@chromium.org

Change-Id: I0803b07a06db9bb10c929ca1444bb08ec6fa9300
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/3906092
Reviewed-by: Jonah Hooper <jonahhooper@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>