Add @project realm.
This realm will be a place where resources which are global to the whole
project can reside.

We considered using @root for this, but since permissions bound in @root
implicitly apply to all other realms in the project, we thought it would
be better to avoid this.

@project has no special meaning, other than being a reserved realm name
with the same meaning across all projects.

R=vadimsh

Bug: 1359069
Change-Id: Ia834ece44408c47392003534feb8ac403b9c26ec
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/3870026
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>