[swarming] Refactor the bot poll loop code.
This is in preparation for adding RBE session management there.
It will be easier to do if this code looks more like a state
machine.

The largest change is moving errors retry loop from poll(...)
guts into the outer poll loop. This allows to do stuff in between
attempts to contact Swarming server (and also makes super explicit
the retry behavior of the bot). This will be important, since we
need to keep pinging RBE session to keep it alive, and thus can't
afford poll(...) getting stuck in the retry loop for long.

Other changes:
  * Decouple hooks calls from poll(...) response handling. Most hooks
    will also be called when polling RBE session.
  * Stop adding "/bot/poll" duration to USD cost of a task. It
    doesn't add much value, just complicates code.
  * Consistently catch interrupts when waiting on an event.

R=iannucci@chromium.org, jonahhooper@google.com
BUG=1377118

Change-Id: I50291d29b0fd48b2c743b13f51025d51576707be
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/3995424
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>