[components] Add a helper for fetching and caching GSM secrets.
Will be used by Swarming to fetch a secret used to HMAC PollState
messages. The Python portion of the code will only be using it
to produce HMAC tags (never to verify them), so it will need only
one active version of the secret (unlike the Go portion which needs
to know a previous and future versions to support graceful
rotation, it is already implemented there).

R=iannucci@chromium.org, jonahhooper@google.com
BUG=1377118

Change-Id: I665a49361d15f91eff55b119b3567ae71ca2a0c7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/4075015
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Jonah Hooper <jonahhooper@google.com>