[swarming] Refactor schedule_request(...) a bit.
This exposes the actual datastore transaction that stores TaskRequest
and other entities. We'll need to transactionally enqueue a TQ task
there. It was not possible when using datastore_utils.insert.

Also datastore_utils.insert has convoluted API and is prone to
submitting duplicate entities in case of transient transaction
errors (where the transaction actually lands, but the client
doesn't know about it and retries).

R=iannucci@chromium.org, jonahhooper@google.com
BUG=1377118

Change-Id: I1f321734f44fa248d6ff9a4485180417ce765355
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/4251687
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>