[swarming] Send keep-alive pings to RBE when running the task.
They can also theoretically be used to detect when the lease
is canceled, but we'll need to cancel through Swarming anyway for
now and so can keep using Swarming for that.

The primary value in these pings is to keep RBE informed there
are bots alive out there with a particular set of dimensions (and
it is OK to keep tasks that match them queued up). Without pings
RBE would consider a bot running a long task to be dead and it
will reject tasks that match this bot with "no matching bots"
error.

R=iannucci@chromium.org, jonahhooper@google.com
BUG=1377118

Change-Id: I38a4db6bcf0e557ee4c703c130e52e0f384865d9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/4285242
Reviewed-by: Jonah Hooper <jonahhooper@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>