Add Windows display_attached dimension
Adds the display_attached dimension to Windows machines. This already
exists for Mac machines, but is being added for Windows since we have
found an approach for Windows that appears to do what we want.

This dimension will be 1 when it can be confirmed that there is a
valid display/monitor attached to the machine, otherwise 0.

Bug: 1194837
Change-Id: I32f3882a97303a113f042156539e64c1b5b61054
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/4375342
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>