[Win][ARM64] fix cpu identifier regex
The regex used for parsing the CPU Identifier was breaking on whitespace
mismatches and the Model value. I have adjusted the regex to be more
flexible with whitepace and updated the Model value match to align with
revision's value. This has worked in our local testing on a windows
ARM64 bot.

Bug:b/260903737
Change-Id: Ided0ea1f86e2415d52414de8a2d332247825bad7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/4377901
Auto-Submit: Adam Todd <actodd@google.com>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Adam Todd <actodd@google.com>