[swarming] Add RBESession.recreate method.
It "reopens" a dead session using latest known parameters. It will
be used by bots in hybrid mode to revive session that die while
the bot executes a long Swarming task. In the "pure" RBE mode such
session are revived in the main bot loop already, but this loop is
on pause while a task is being executed. Instead there's another
mini-loop inside task_runner which will use RBESession.recreate()
to reopen dead sessions.

This is particularly important while we abuse UNHEALTHY status
as a replacement for not-yet-existing MAINTENANCE status: RBE
scheduler kills sessions with UNHEALTHY status after couple of
minutes.

But it won't hurt even if we use MAINTENANCE. Session should not
be dying in this case, but if they do, for whatever reason, we
still need to reopen them.

R=jonahhooper@google.com
BUG=1377118

Change-Id: I1756be1e8cc27bcc5d75377619067f9d79446d18
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/4571735
Reviewed-by: Jonah Hooper <jonahhooper@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>