[swarming] prpc TaskResultResponse - changed exit_code to int64
This changes `exit_code` to `int64` to deal with - https://chromium-swarm.appspot.com/bot?id=luci-chromium-try-uca-win-rel-compilator-89-11-qj5t where windows returns a 64bit exit code!

This change is wirecompatible with existing `int32` values according too - https://stackoverflow.com/questions/48987379/how-to-update-a-running-system-using-protobuf-from-int32-to-64

It will be backwards incompatible if the server is sending 64bit values to clients which expect 32bit. Since doing that now would 500 then the only instance of this being an issue is that which we observed on frontend so I think this change should be relatively safe.

Bug: 1410932
Change-Id: I1d86f92705e4960f6779d73eca6706e8f345850f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/4582967
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Auto-Submit: Jonah Hooper <jonahhooper@google.com>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>