[swarming] Reduce latency of reporting bots as idle.
Apparently Skia task scheduler periodically lists Swarming bots
and submits tasks only when it notices a bot is idle. It means
reporting the bot as idle ASAP is important to keep Skia's
scheduling latency low.

Make the first RBE poll after running a task non-blocking. If it
returns no leases, we'll report the bot as idle to Swarming ASAP
and subsequent polls will be blocking. And if returns a lease,
we'll just execute it as usual (no need to report the bot as
idle in this case, because it isn't).

Also reduce the sleeping time in one loop iteration in the
hybrid mode 2x. Since we alternate polling from Swarming and RBE,
the full polling cycles is two iterations. We sleep twice, and
need to cut the individual sleeping time accordingly to match
pure Swarming bot sleeping behavior in the end.

R=jonahhooper@google.com
BUG=1377118

Change-Id: Ib80df31c835dc7b06ac965b419ec18af25c7bf67
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/4604505
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Jonah Hooper <jonahhooper@google.com>