[swarming] Report lease result ASAP after task completion.
Swarming runs user supplied hooks and various cleanups after
completing a task and before polling for the next task.

RBE API normally combines reporting a task result with picking up
the next task into a single UpdateBotSession(...) call. It means
if hooks and cleanups run for too long, the result report is
delayed as well. This seemed to cause some leases to fail with
LOST status, because RBE thinks the bot abandons them, because it
stops pinging them.

Now that we have MAINTENANCE status, we can send UpdateBotSession
call that just reports the result, but doesn't pick up any new
leases. We can do it as soon as the task finishes. This should
reduce the number of LOST leases.

Note that if the cleanup and hooks still take too long to run, the
bot session itself will die by timeout, but this timeout seems
to be much higher than the timeout for LOST leases and the bot in
theory is already capable of reopening the session.

Also reorder some calls to make sure Swarming always "closes"
the failed task before RBE is notified the lease is done.
Otherwise RBE pubsub may end up racing with _post_error_task,
which can be potentially bad.

R=jonahhooper@google.com

Change-Id: I02ce7c312a128535cbde2fd3e727e846b3a9b046
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/4702676
Reviewed-by: Jonah Hooper <jonahhooper@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>