[swarming] Add support for detecting ssd type 'nvme_card'
The additional SSD added to Windows host is not being detected as MediaType 4 (which is an SSD). See: https://learn.microsoft.com/en-us/windows-hardware/drivers/storage/msft-physicaldisk
Instead this is seen as MediaType 0 (which is Unspecified)
As a result the SSD dimension will report false.

This is a change to add detection of the Model if it equals 'nvme_card' so we can report true back on the SSD dimension. There are other properties we can key off of rather than model, however this seems the most appropriate. See: https://paste.googleplex.com/5847487439962112

Bug:b/288623226
Change-Id: I251bcedb9c372270f8fe085b0dfe5b670cbc6eef
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/4727291
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Sam Dockery <samdockery@google.com>