[swarming] Increase memcache claim timeout 15s => 60s.
This is used to coordinate between bots that want to
consume the same task. Only on of them gets the claim,
which allows it to do a datastore transaction to actually
consume the task.

Looks like sometimes this transaction takes a lot of time,
and the claim expires, which leads to other bots trying to
claim the task concurrently.

Increase the claim deadline to workaround that. Note that this
will likely increase scheduling tail latency (for cases when
a bot claims the task, but then fails to land the transaction).

R=jonahhooper@google.com

Change-Id: I5fef06e3cc6e2c510af327485978bcbad9f37bb6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/4864398
Reviewed-by: Jonah Hooper <jonahhooper@google.com>
Reviewed-by: Sergey Fetisov <sfetisov@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>