[swarming] Allow no session token in /rbe/session/update response.
This can happen if the bot was stuck for a long time and its
existing session token has expired and cannot longer be
refreshed.

The server will treat this situation in the same way as the
RBE session expiration (because the RBE session expires too),
in particular it will return the session in BOT_TERMINATING state.

The Swarming bot already knows how to handle this (by recreating
the session), the only problem is that it currently *always*
expects a valid session token in /rbe/session/update response.

This CL relaxes this requirement.

R=jonahhooper@google.com, sfetisov@google.com
BUG=1377118

Change-Id: Ie5cd3176a8b409bb771fd1cb80501c534b4a8531
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/4898402
Reviewed-by: Sergey Fetisov <sfetisov@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>