Make utils.to_json_encodable gracefully degrade when protorpc is missing.
The majority of value of to_json_encodable is to be kind of
a "anything-in, jsonish-value-out". For programs which use protorpc,
protorpc Message objects may be passed to this function.

However, protorpc is very ancient, and very deprecated. This CL
makes it so that uses of `utils` will not immediately fail in
environments which don't have protorpc available.

The import chain that led here was:

    prpc.client -> net -> auth -> utils

R=dbeckett@google.com, mkiedys@google.com, vadimsh

Change-Id: Ief536bfa945812be9805a670ec8351837048e622
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/4950736
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: Derek Beckett <dbeckett@chromium.org>