[auth-service] Add permissionslist to AuthRealmsGlobals model
During the test of updating permissions.cfg and checking if
AuthRealmsGlobals was subsequently updated, the transaction failed due
to the AuthRealmsGlobals model not having the attribute
`permissionslist`. This field is populated by the Go version of Auth
Service.

This CL adds the field `permissionslist` to the AuthRealmsGlobals model,
so that making a historical copy will succeed.

Bug: b:311110438
Change-Id: Iac9bf9c58e24583acbb569dbfb8dfbf09e032998
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/5051646
Reviewed-by: Joey Scarr <jsca@google.com>
Commit-Queue: Anne Redulla <aredulla@google.com>