[swarming] Route more pubsub notification tasks to Go
Tested this feature with Buildbucket on both Dev and Prod which worked
well. We can enable for more. In Swarming Dev, there are 2 customers (BB
and skia-buildbots-stage). In Swarming Prod, there are 3 customers (BB,
skia-buildbots and pinpoint). This CL enables 100% on Dev. For BB, it
increase 25% -> 100% on Prod.


Bug:b/325342884
Change-Id: Ib7e6f9754279a163a90c7842720a9e62219e9d0b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/5403219
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Yuanjun Huang <yuanjunh@google.com>