[swarming] Report to monitoring why task slices expire.
This will be used to confirm that RBE's PubSub notifications
are 100% reliable and we can get rid of the cron job that
scans datastore.

Note that PubSub can have a bit of latency, so we will also need
to change the cron job to expire tasks after some delay to give
PubSub a chance to do it first (otherwise occasionally the cron job
will be winning the race and we'll have unclean monitoring signal).
This will be done in a separate CL.

R=chanli@chromium.org

Change-Id: I358f4e4f925374b65f4869bbfe498ec3ecfa98c8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/5607981
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>