[swarming] Update "Last Known Good Bot Code" faster.
"Last Known Good Bot Code" is a version of the bot code
that is launched when the bot machine is rebooted. It
exists as a safeguard mechanism in case a new bot code is
so broken, that it can't self-update itself to a non-broken
version. Theoretically, if there's such super-broken release,
we'll need to reboot all machines to make them launch
non-broken "last known good bot code".

Before this CL, "Last Known Good Bot Code" was updated to
be the currently running bot code either after a successful
task or after **1 week** of idleness or failing tasks.

There are many bots that are idle or execute only failing tasks.
Every time such machine is rebooted (either explicitly or by
the Swarming bot itself), it launches up to 1w old bot executable,
which then promptly self-updates itself to the current version.
But this self-update code is up to 1w old.

It means when changing the bot code we need to wait at least
a week for all bots to permanently pick up the change and for
the old code to be completely flushed out.

This is very inconvenient during refactoring of the bot
deployment mechanism. It is also overly cautious. As long
as we can execute self-update, we should be good.

So update LKGBC either after a task (successful or not)
or after an idle poll cycle.

R=chanli@chromium.org
BUG=b/355013257

Change-Id: I1962f2404d228ae5db0df1675c1e7db1a389b1f5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/5813789
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>