Add Mac thermal state
Adds the ability to retrieve the thermal state on Mac, which is a coarse
way of determining whether a machine is experiencing cooling issues or
not. This will ultimately be used to auto-quarantine Mac devices which
are too hot to run tests.

Bug: 325055849
Change-Id: I28842ddc523503bcf29ae6c70c7c0e3ac65eb173
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/5921863
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Jonathan Pittman <jmpittman@google.com>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>