[swarming] Add hooks script SHA256 to BotGroupConfig tuple.
This will be used when generating `handshake_config_hash` in
the session token and deciding if the bot should be restarted
to pick up a new config. Store it to avoid recalculating this
hash on every /bot/poll (the bot hooks file can be pretty
large).

Also remove logging of binary and/or huge protobuf protos, it's
not helpful (and this code is hopefully will be deleted soon
anyway).

Also run "git cl format".

R=chanli@chromium.org
BUG=b/355013282

Change-Id: I1b0d10e1e8565cfd5d3b204bf32e5ae0e8c93c5f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/6016148
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>