[Swarming] Py: Add string_pair_serializer
In https://crrev.com/c/6112782 I plan to use a different way to calculate the properties hashes on the go side.

To ensure the deduplication by properties hash work during the migration, we need to calculate the same properties_hash from the py side as the go side.

This CL implements the same string_pair_serializer as go side to serialize properties, and I plan to use it in a follow-up CL to calculate properties hash.

Bug: 355013435
Change-Id: Idfe181bad519855620947c12da7c87ece1c8c1a5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/6124270
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Chan Li <chanli@chromium.org>