[swarming] Refuse running migrated bot API handlers in prod.
Ripping them out completely is a lot of effort (fixing or removing
convoluted tests) with dubious impact. It will likely be simpler
to just delete everything at once in the end.

Maintaining them up-to-date with the Go implementation is also
a waste of time. So just keep them as is in their stale form,
but refuse to run them outside of tests.

R=chanli@chromium.org

Change-Id: I5466d5621390d3ab8d9d8ccc1f7500eb7b286c79
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/6266047
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>