[Swarming] bot task update: add `canceled` to ACCEPTED_KEYS
`canceled` is set at http://shortn/_jtJLDRSPbZ for tasks that are canceled at set up stage. The intention was to inform server to mark those tasks as "CANCELED" instead of "KILLED". But it was not added to ACCEPTED_KEYS causing the code path was never reached and a bunch of bot task update failures.

Change-Id: I6aac091369aa4e7de3b07fa44241066761d2d692
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/6384636
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Chan Li <chanli@chromium.org>