Fix swarming's dimensions for Android devices running trunk builds
When running a trunk build, an android device will report "MAIN" for
its "build.id" prop, which isn't very helpful as a swarming dimension.
So this detects that case (generalized from the "if wembley_2GB"
conditional) and appends the incremental build ID and codename.

This also adds another location for the "fingerprint" build prop to
add to the state dict.

Bug: 368004900
Change-Id: I7fb6b75b234ca3d5a5c40064c6a055e895208511
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/6557666
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>